feat(HIGH-14): organism폴더 마이그레이션
