+++ src/components/landing/organism/AnimatedLogo.tsx
@@ -1,0 +1,16 @@
+import { AnimatePresence } from "framer-motion";
+import AnimatedImageEntrance from "@/components/common/atom/AnimatedImageEntrance";
+
+interface AnimatedLogoProps {
+  isVisible: boolean;
+  src: string;
+  alt: string;
+}
+
+const AnimatedLogo = ({ isVisible, src, alt }: AnimatedLogoProps) => (
+  <AnimatePresence>
+    {isVisible && <AnimatedImageEntrance src={src} alt={alt} />}
+  </AnimatePresence>
+);
+
+export default AnimatedLogo;
